did changes controller and cors

+++ src/main/java/com/bookverser/BookVerse/dto/OrderResponseDto.java
@@ -6,13 +6,15 @@ import java.util.List;
 
 import com.bookverser.BookVerse.dto.PlaceOrderRequest.AddressRequest;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderResponseDto {
     private Long orderId;
     private Long customerId;
     private String paymentMethod;
     private String status;
-    private double totalAmount;
+    private BigDecimal totalAmount;
     private List<CartItemDto> items;
     private AddressResponseDto shippingAddress;
 }
